# Conflicts:
#	dss-asic-xades/src/main/java/eu/europa/esig/dss/asic/xades/signature/ASiCWithXAdESService.java

+++ dss-pades-openpdf/src/main/java/eu/europa/esig/dss/pdf/openpdf/visible/ImageOnlySignatureDrawer.java
@@ -70,7 +70,7 @@ public class ImageOnlySignatureDrawer extends AbstractITextSignatureDrawer {
 		image.setAbsolutePosition(x, y);
 		image.scaleAbsolute(width, height);
 
-		image.setRotationDegrees((float) (ImageRotationUtils.ANGLE_360 - finalRotation)); // opposite rotation
+		image.setRotationDegrees(ImageRotationUtils.ANGLE_360 - finalRotation); // opposite rotation
 
 		PdfTemplate layer = appearance.getLayer(2);
 		Rectangle boundingBox = layer.getBoundingBox();
